+++ ops-lib/src/model/task.rs
@@ -1,4 +1,4 @@
-use crate::ops_model::status::Status;
+use crate::model::status::Status;
 use serde::{Deserialize, Serialize};
 use serde_json::{Map, Value};